requirements.txt
customtkinter
pygame
arabic_reshaper
python-bidi
requests

